impls/deref.rs
